\name{randnet-package}
\alias{randnet-package}
\alias{randnet}
\docType{package}
\title{
Statistical modeling of random networks with model selection and
parameter tuning
}
\description{
The package provides model fitting and estimation functions for some
popular random network models. More importantly, it implements a general
cross-validation framework for model selection and parameter tuning
called ECV. Several other model selection methods are also included.
}
\details{
\tabular{ll}{
Package: \tab randnet\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2017-09-13\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
  Tianxi Li, Elizaveta Levina, Ji Zhu\cr

Maintainer: Tianxi Li <tianxili@umich.edu>

}
\references{
  T. Li, E. Levina, and J. Zhu. Network cross-validation by edge sampling. arXiv preprint arXiv:1612.04717, 2016.
}
\keyword{ package }
